crypto/ex3/pyproject.toml
[tool.poetry]
name = "ex3"
version = "0.1.0"
description = ""
authors = ["memetelve <memetelve1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
